Split server-users.test.ts (1,244 lines) into 6 focused files
Replaces the monolithic test file with targeted files, each <=300 lines:
- server-users-db.test.ts: CRUD, invited users, isInviteExpired, db edge cases
- server-users-routes-list-and-create.test.ts: GET/POST /admin/users listing and invite creation
- server-users-routes-mutations.test.ts: delete and activate user routes
- server-users-auth-and-join.test.ts: login flow, join flow, CSRF/expired invite edge cases
- server-users-permissions-and-session.test.ts: role enforcement, session, settings, utils edge cases
- server-users-ui-and-observability.test.ts: navigation, setup page, template rendering, audit logging

No tests were dropped or renamed. All suites still run under describeWithEnv(..., { db: true }, ...).
