--- test/lib/server-users.test.ts
@@ -1,1244 +1,0 @@
-import { expect } from "@std/expect";
-import { describe, it as test } from "@std/testing/bdd";
-import { getSessionCookieName } from "#lib/cookies.ts";
-import { encrypt } from "#lib/crypto/encryption.ts";
-import { hashPassword } from "#lib/crypto/hashing.ts";
-import { getAllActivityLog } from "#lib/db/activityLog.ts";
-import { getDb, insert } from "#lib/db/client.ts";
-import { createSession } from "#lib/db/sessions.ts";
-import {
-  createInvitedUser,
-  decryptAdminLevel,
-  decryptUsername,
-  getAllUsers,
-  getUserByUsername,
-  hasPassword,
-  invalidateUsersCache,
-  isInviteExpired,
-  isInviteValid,
-  verifyUserPassword,
-} from "#lib/db/users.ts";
-import { handleRequest } from "#routes";
-import {
-  adminFormPost,
-  assertAdminHtmlWithCookie,
-  assertPublicHtml,
-  awaitTestRequest,
-  createTestInvite,
-  createTestManagerSession,
-  describeWithEnv,
-  expectHtmlResponse,
-  expectRedirect,
-  expectRedirectWithFlash,
-  FLASH_TEST_ID,
-  flashCookieHeader,
-  mockAdminLoginRequest,
-  mockFormRequest,
-  mockRequest,
-  submitJoinForm,
-  TEST_ADMIN_PASSWORD,
-  TEST_ADMIN_USERNAME,
-  testCookie,
-  testRequiresAuth,
-} from "#test-utils";
-
-describeWithEnv("server (multi-user admin)", { db: true }, () => {
-  describe("users CRUD", () => {
-    test("createTestDbWithSetup creates the owner user", async () => {
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      expect(user!.id).toBe(1);
-      expect(user!.wrapped_data_key).not.toBeNull();
-
-      const level = await decryptAdminLevel(user!);
-      expect(level).toBe("owner");
-
-      const username = await decryptUsername(user!);
-      expect(username).toBe(TEST_ADMIN_USERNAME);
-    });
-
-    test("verifyUserPassword returns hash for correct password", async () => {
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      const hash = await verifyUserPassword(user!, TEST_ADMIN_PASSWORD);
-      expect(hash).toBeTruthy();
-      expect(hash).toContain("pbkdf2:");
-    });
-
-    test("verifyUserPassword returns null for wrong password", async () => {
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(user).not.toBeNull();
-      const result = await verifyUserPassword(user!, "wrongpassword");
-      expect(result).toBeNull();
-    });
-
-    test("getUserByUsername returns null for nonexistent user", async () => {
-      const user = await getUserByUsername("nonexistent");
-      expect(user).toBeNull();
-    });
-
-    test("getAllUsers returns all users", async () => {
-      const users = await getAllUsers();
-      expect(users.length).toBe(1);
-      expect(users[0]!.id).toBe(1);
-    });
-  });
-
-  describe("invited users", () => {
-    test("createInvitedUser creates user with invite code", async () => {
-      const inviteHash = await hashPassword("invite123");
-      const expiry = new Date(Date.now() + 86400000).toISOString();
-
-      const user = await createInvitedUser(
-        "invitee",
-        "manager",
-        inviteHash,
-        expiry,
-      );
-
-      expect(user.id).toBe(2);
-      expect(user.password_hash).toBe("");
-      expect(user.wrapped_data_key).toBeNull();
-
-      const level = await decryptAdminLevel(user);
-      expect(level).toBe("manager");
-
-      const username = await decryptUsername(user);
-      expect(username).toBe("invitee");
-
-      const hasPwd = await hasPassword(user);
-      expect(hasPwd).toBe(false);
-    });
-
-    test("isInviteValid returns true for valid invite", async () => {
-      const expiry = new Date(Date.now() + 86400000).toISOString();
-      const user = await createInvitedUser(
-        "invitee",
-        "manager",
-        "somehash",
-        expiry,
-      );
-
-      const valid = await isInviteValid(user);
-      expect(valid).toBe(true);
-    });
-
-    test("isInviteValid returns false for expired invite", async () => {
-      const expiry = new Date(Date.now() - 1000).toISOString();
-      const user = await createInvitedUser(
-        "expired-user",
-        "manager",
-        "somehash",
-        expiry,
-      );
-
-      const valid = await isInviteValid(user);
-      expect(valid).toBe(false);
-    });
-  });
-
-  describe("isInviteExpired", () => {
-    test("returns true for expired invite", async () => {
-      const expiry = new Date(Date.now() - 1000).toISOString();
-      const user = await createInvitedUser(
-        "expired-check",
-        "manager",
-        "somehash",
-        expiry,
-      );
-
-      expect(await isInviteExpired(user)).toBe(true);
-    });
-
-    test("returns false for valid invite", async () => {
-      const expiry = new Date(Date.now() + 86400000).toISOString();
-      const user = await createInvitedUser(
-        "valid-check",
-        "manager",
-        "somehash",
-        expiry,
-      );
-
-      expect(await isInviteExpired(user)).toBe(false);
-    });
-
-    test("returns false for user without invite code", async () => {
-      const owner = await getUserByUsername(TEST_ADMIN_USERNAME);
-      expect(await isInviteExpired(owner!)).toBe(false);
-    });
-  });
-
-  describe("role enforcement", () => {
-    test("manager user cannot access settings page", async () => {
-      // Create a manager user with a password
-      const hash = await hashPassword("managerpass");
-      const encHash = await encrypt(hash);
-      await getDb().execute(
-        insert("users", {
-          admin_level: await encrypt("manager"),
-          password_hash: encHash,
-          username_hash: await encrypt("manager"),
-          username_index: "manager-idx-unique",
-          // Give the manager a wrapped_data_key so login works
-          wrapped_data_key: (await getUserByUsername(TEST_ADMIN_USERNAME))!
-            .wrapped_data_key,
-        }),
-      );
-      invalidateUsersCache();
-
-      // Create a session for the manager user
-      const managerUserId = 2;
-      await createSession(
-        "manager-token",
-        "manager-csrf",
-        Date.now() + 3600000,
-        null,
-        managerUserId,
-      );
-
-      // Manager should get 403 on owner-only routes
-      const settingsResponse = await awaitTestRequest("/admin/settings", {
-        cookie: `${getSessionCookieName()}=manager-token`,
-      });
-      expect(settingsResponse.status).toBe(403);
-
-      const sessionsResponse = await awaitTestRequest("/admin/sessions", {
-        cookie: `${getSessionCookieName()}=manager-token`,
-      });
-      expect(sessionsResponse.status).toBe(403);
-
-      const usersResponse = await awaitTestRequest("/admin/users", {
-        cookie: `${getSessionCookieName()}=manager-token`,
-      });
-      expect(usersResponse.status).toBe(403);
-    });
-
-    test("owner user can access settings page", async () => {
-      const response = await awaitTestRequest("/admin/settings", {
-        cookie: await testCookie(),
-      });
-      expect(response.status).toBe(200);
-    });
-
-    test("owner user can access sessions page", async () => {
-      const response = await awaitTestRequest("/admin/sessions", {
-        cookie: await testCookie(),
-      });
-      expect(response.status).toBe(200);
-    });
-
-    test("owner user can access users page", async () => {
-      const response = await awaitTestRequest("/admin/users", {
-        cookie: await testCookie(),
-      });
-      await expectHtmlResponse(response, 200, "Users", TEST_ADMIN_USERNAME);
-    });
-
-    test("manager user can access dashboard", async () => {
-      const cookie = await createTestManagerSession(
-        "mgr-dash-session",
-        "dashmanager",
-      );
-      await assertAdminHtmlWithCookie("/admin/", cookie, "Events");
-    });
-  });
-
-  describe("GET /admin/users", () => {
-    testRequiresAuth("/admin/users");
-
-    test("shows users list when authenticated as owner", async () => {
-      const response = await awaitTestRequest("/admin/users", {
-        cookie: await testCookie(),
-      });
-      await expectHtmlResponse(response, 200, TEST_ADMIN_USERNAME, "owner");
-    });
-  });
-
-  describe("GET /admin/users (with query params)", () => {
-    test("displays invite link from query param", async () => {
-      const response = await awaitTestRequest(
-        "/admin/users?invite=" +
-          encodeURIComponent("https://localhost/join/abc123"),
-        { cookie: await testCookie() },
-      );
-      await expectHtmlResponse(
-        response,
-        200,
-        "https://localhost/join/abc123",
-        "Invite link",
-      );
-    });
-
-    test("displays success message from flash cookie", async () => {
-      const cookie = await testCookie();
-      const response = await awaitTestRequest(
-        `/admin/users?flash=${FLASH_TEST_ID}`,
-        {
-          cookie: `${cookie}; ${flashCookieHeader(
-            "User deleted successfully",
-          )}`,
-        },
-      );
-      await expectHtmlResponse(
-        response,
-        200,
-        "User deleted successfully",
-        'class="success"',
-      );
-    });
-  });
-
-  describe("GET /admin/user/new", () => {
-    testRequiresAuth("/admin/user/new");
-
-    test("renders invite user form when authenticated as owner", async () => {
-      const response = await awaitTestRequest("/admin/user/new", {
-        cookie: await testCookie(),
-      });
-      await expectHtmlResponse(
-        response,
-        200,
-        "Invite User",
-        'action="/admin/users"',
-      );
-    });
-  });
-
-  describe("POST /admin/users (invite)", () => {
-    testRequiresAuth("/admin/users", {
-      body: {
-        admin_level: "manager",
-        username: "newuser",
-      },
-      method: "POST",
-    });
-
-    test("creates invited user and shows invite link", async () => {
-      const { response } = await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "newmanager",
-      });
-
-      expect(response.status).toBe(302);
-      const location = expectRedirect(response);
-      expect(decodeURIComponent(location)).toContain("/join/");
-
-      // Verify user was created in the database
-      const users = await getAllUsers();
-      expect(users.length).toBe(2);
-    });
-
-    test("rejects duplicate username", async () => {
-      const { response } = await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: TEST_ADMIN_USERNAME,
-      });
-
-      expectRedirectWithFlash(
-        "/admin/user/new",
-        expect.stringContaining("already taken"),
-        false,
-      )(response);
-    });
-
-    test("rejects invalid role", async () => {
-      const { response } = await adminFormPost("/admin/users", {
-        admin_level: "superadmin",
-        username: "newuser",
-      });
-
-      expectRedirectWithFlash(
-        "/admin/user/new",
-        expect.stringContaining("Invalid role"),
-        false,
-      )(response);
-    });
-  });
-
-  describe("GET /admin/users/:id/delete", () => {
-    test("shows delete confirmation page", async () => {
-      // Create a user to delete
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "todelete",
-      });
-
-      const response = await awaitTestRequest("/admin/users/2/delete", {
-        cookie: await testCookie(),
-      });
-      await expectHtmlResponse(
-        response,
-        200,
-        "Delete User",
-        "todelete",
-        "confirm_identifier",
-      );
-    });
-
-    test("returns 404 for nonexistent user", async () => {
-      const response = await awaitTestRequest("/admin/users/999/delete", {
-        cookie: await testCookie(),
-      });
-      expect(response.status).toBe(404);
-    });
-
-    test("rejects deleting self", async () => {
-      const response = await awaitTestRequest("/admin/users/1/delete", {
-        cookie: await testCookie(),
-      });
-      await expectHtmlResponse(response, 400, "Cannot delete your own account");
-    });
-  });
-
-  describe("POST /admin/users/:id/delete", () => {
-    test("deletes a user with correct confirmation", async () => {
-      // Create an invited user first
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "deleteme",
-      });
-
-      const usersBefore = await getAllUsers();
-      expect(usersBefore.length).toBe(2);
-
-      // Delete user with id 2 with correct confirmation
-      const { response } = await adminFormPost("/admin/users/2/delete", {
-        confirm_identifier: "deleteme",
-      });
-      expectRedirectWithFlash(
-        "/admin/users",
-        expect.stringContaining("deleted"),
-      )(response);
-
-      const usersAfter = await getAllUsers();
-      expect(usersAfter.length).toBe(1);
-    });
-
-    test("rejects deletion with wrong confirmation", async () => {
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "keepme",
-      });
-
-      const { response } = await adminFormPost("/admin/users/2/delete", {
-        confirm_identifier: "wrongname",
-      });
-      expectRedirectWithFlash(
-        "/admin/users/2/delete",
-        expect.stringContaining("Username does not match"),
-        false,
-      )(response);
-
-      const usersAfter = await getAllUsers();
-      expect(usersAfter.length).toBe(2);
-    });
-
-    test("rejects deletion without confirmation", async () => {
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "keepme2",
-      });
-
-      const { response } = await adminFormPost("/admin/users/2/delete");
-      expectRedirectWithFlash(
-        "/admin/users/2/delete",
-        expect.stringContaining("Username does not match"),
-        false,
-      )(response);
-    });
-
-    test("prevents deleting self", async () => {
-      const { response } = await adminFormPost("/admin/users/1/delete", {
-        confirm_identifier: TEST_ADMIN_USERNAME,
-      });
-      await expectHtmlResponse(response, 400, "Cannot delete your own account");
-    });
-
-    test("deletes another owner with correct confirmation", async () => {
-      // Create another owner user
-      await adminFormPost("/admin/users", {
-        admin_level: "owner",
-        username: "otheradmin",
-      });
-
-      const usersBefore = await getAllUsers();
-      expect(usersBefore.length).toBe(2);
-
-      // Delete the other owner with correct confirmation
-      const { response } = await adminFormPost("/admin/users/2/delete", {
-        confirm_identifier: "otheradmin",
-      });
-      expectRedirectWithFlash(
-        "/admin/users",
-        expect.stringContaining("deleted"),
-      )(response);
-
-      const usersAfter = await getAllUsers();
-      expect(usersAfter.length).toBe(1);
-    });
-  });
-
-  describe("login flow", () => {
-    test("login with username and password", async () => {
-      const response = await handleRequest(
-        await mockAdminLoginRequest({
-          password: TEST_ADMIN_PASSWORD,
-          username: TEST_ADMIN_USERNAME,
-        }),
-      );
-      expectRedirectWithFlash("/admin", "Logged in")(response);
-      const sessionCookie = response.headers
-        .getSetCookie()
-        .find((c) => c.startsWith(`${getSessionCookieName()}=`));
-      expect(sessionCookie).toBeDefined();
-    });
-
-    test("login with wrong username returns 401", async () => {
-      const response = await handleRequest(
-        await mockAdminLoginRequest({
-          password: TEST_ADMIN_PASSWORD,
-          username: "nonexistent",
-        }),
-      );
-      expectRedirectWithFlash(
-        "/admin",
-        expect.stringContaining("Username or password was wrong"),
-        false,
-      )(response);
-    });
-
-    test("login with wrong password returns 401", async () => {
-      const response = await handleRequest(
-        await mockAdminLoginRequest({
-          password: "wrongpassword",
-          username: TEST_ADMIN_USERNAME,
-        }),
-      );
-      expectRedirectWithFlash(
-        "/admin",
-        expect.stringContaining("Username or password was wrong"),
-        false,
-      )(response);
-    });
-
-    test("login page shows username field", async () => {
-      await assertPublicHtml("/admin/", "username");
-    });
-  });
-
-  describe("join flow", () => {
-    test("GET /join/:code returns 404 for invalid code", async () => {
-      const response = await handleRequest(mockRequest("/join/invalidcode123"));
-      await expectHtmlResponse(response, 404, "invalid");
-    });
-
-    test("GET /join/:code returns join page for valid invite", async () => {
-      const { inviteCode } = await createTestInvite("joiner");
-
-      const joinResponse = await handleRequest(
-        mockRequest(`/join/${inviteCode}`),
-      );
-      expect(joinResponse.status).toBe(200);
-      const joinHtml = await joinResponse.text();
-      expect(joinHtml).toContain("joiner");
-      expect(joinHtml).toContain("password");
-    });
-
-    test("GET /join/complete shows confirmation page", async () => {
-      await assertPublicHtml("/join/complete", "Password Set");
-    });
-
-    test("POST /join/:code sets password for invited user", async () => {
-      const { inviteCode } = await createTestInvite("joiner2");
-
-      const joinPostResponse = await submitJoinForm(inviteCode, {
-        password: "newpassword123",
-        password_confirm: "newpassword123",
-      });
-
-      expectRedirectWithFlash(
-        "/join/complete",
-        "Password set successfully",
-      )(joinPostResponse);
-
-      // Verify user now has a password
-      const user = await getUserByUsername("joiner2");
-      expect(user).not.toBeNull();
-      const hasPwd = await hasPassword(user!);
-      expect(hasPwd).toBe(true);
-    });
-
-    test("POST /join/:code rejects mismatched passwords", async () => {
-      const { inviteCode } = await createTestInvite("joiner3");
-
-      const joinPostResponse = await submitJoinForm(inviteCode, {
-        password: "newpassword123",
-        password_confirm: "differentpassword",
-      });
-
-      expectRedirectWithFlash(
-        `/join/${inviteCode}`,
-        expect.stringContaining("do not match"),
-        false,
-      )(joinPostResponse);
-    });
-
-    test("POST /join/:code rejects short passwords", async () => {
-      const { inviteCode } = await createTestInvite("joiner4");
-
-      const joinPostResponse = await submitJoinForm(inviteCode, {
-        password: "short",
-        password_confirm: "short",
-      });
-
-      expectRedirectWithFlash(
-        `/join/${inviteCode}`,
-        expect.stringContaining("8 characters"),
-        false,
-      )(joinPostResponse);
-    });
-  });
-
-  describe("navigation", () => {
-    test("owner sees all nav links", async () => {
-      const response = await awaitTestRequest("/admin/", {
-        cookie: await testCookie(),
-      });
-      const html = await response.text();
-      expect(html).toContain("Settings");
-      expect(html).toContain("Users");
-    });
-
-    test("manager does not see owner-only nav links", async () => {
-      const cookie = await createTestManagerSession(
-        "navmgr-session",
-        "navmanager",
-      );
-      const dashboardResponse = await awaitTestRequest("/admin/", { cookie });
-      const html = await dashboardResponse.text();
-      expect(html).not.toContain("Settings");
-      expect(html).not.toContain("Users");
-    });
-
-    test("manager sees log nav link", async () => {
-      const cookie = await createTestManagerSession(
-        "navmgr-log-session",
-        "navmanagerlog",
-      );
-      const dashboardResponse = await awaitTestRequest("/admin/", { cookie });
-      const html = await dashboardResponse.text();
-      expect(html).toContain("/admin/log");
-    });
-  });
-
-  describe("setup page", () => {
-    test("setup includes admin_username field", async () => {
-      // Need to start fresh without setup
-      const { getDb: getDbFn } = await import("#lib/db/client.ts");
-      await getDbFn().execute("DELETE FROM settings");
-      await getDbFn().execute("DELETE FROM users");
-      const { settings: s } = await import("#lib/db/settings.ts");
-      s.setup.clearCache();
-      s.invalidateCache();
-      invalidateUsersCache();
-
-      await assertPublicHtml("/setup/", "admin_username");
-    });
-  });
-
-  describe("POST /admin/users/:id/activate", () => {
-    test("activates user who has set password", async () => {
-      const { inviteCode, cookie, csrfToken } =
-        await createTestInvite("activateme");
-
-      await submitJoinForm(inviteCode, {
-        password: "newpassword123",
-        password_confirm: "newpassword123",
-      });
-
-      // Now activate user id 2
-      const activateResponse = await handleRequest(
-        mockFormRequest(
-          "/admin/users/2/activate",
-          { csrf_token: csrfToken },
-          cookie,
-        ),
-      );
-      expectRedirectWithFlash(
-        "/admin/users",
-        expect.stringContaining("activated successfully"),
-      )(activateResponse);
-    });
-
-    test("returns 404 for nonexistent user", async () => {
-      const { response } = await adminFormPost("/admin/users/999/activate");
-      await expectHtmlResponse(response, 404, "User not found");
-    });
-
-    test("rejects user who has not set password", async () => {
-      // Create invite but don't complete join flow
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "nopassword",
-      });
-
-      const { response } = await adminFormPost("/admin/users/2/activate");
-      await expectHtmlResponse(response, 400, "not set their password");
-    });
-
-    test("rejects already activated user", async () => {
-      // User 1 (the owner) is already activated
-      const { response } = await adminFormPost("/admin/users/1/activate");
-      await expectHtmlResponse(response, 400, "already activated");
-    });
-
-    test("returns 500 when session lacks data key", async () => {
-      // Create a session without wrapped_data_key for the owner
-      await createSession(
-        "no-dk-session",
-        "no-dk-csrf",
-        Date.now() + 3600000,
-        null,
-        1,
-      );
-
-      // Create an invited user with password set
-      const { inviteCode } = await createTestInvite("needsactivation");
-      await submitJoinForm(inviteCode, {
-        password: "newpassword123",
-        password_confirm: "newpassword123",
-      });
-
-      // Try to activate using session without data key (need a signed CSRF token)
-      const { signCsrfToken } = await import("#lib/csrf.ts");
-      const csrfToken = await signCsrfToken();
-      const response = await handleRequest(
-        mockFormRequest(
-          "/admin/users/2/activate",
-          { csrf_token: csrfToken },
-          `${getSessionCookieName()}=no-dk-session`,
-        ),
-      );
-      await expectHtmlResponse(response, 500, "session lacks data key");
-    });
-  });
-
-  describe("POST /admin/users/:id/delete (not found)", () => {
-    test("returns 404 for nonexistent user", async () => {
-      const { response } = await adminFormPost("/admin/users/999/delete");
-      await expectHtmlResponse(response, 404, "User not found");
-    });
-  });
-
-  describe("POST /admin/users (form validation)", () => {
-    test("rejects missing username", async () => {
-      const { response } = await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "",
-      });
-      expect(response.status).toBe(302);
-    });
-  });
-
-  describe("join flow (expired invite)", () => {
-    test("GET /join/:code returns 410 for expired invite", async () => {
-      const expiry = new Date(Date.now() - 1000).toISOString();
-      const { hashInviteCode } = await import("#lib/db/users.ts");
-      const codeHash = await hashInviteCode("expired-code-123");
-      await createInvitedUser("expired-join", "manager", codeHash, expiry);
-
-      const response = await handleRequest(
-        mockRequest("/join/expired-code-123"),
-      );
-      await expectHtmlResponse(response, 410, "expired");
-    });
-
-    test("POST /join/:code returns 410 for expired invite", async () => {
-      const expiry = new Date(Date.now() - 1000).toISOString();
-      const { hashInviteCode } = await import("#lib/db/users.ts");
-      const codeHash = await hashInviteCode("expired-post-123");
-      await createInvitedUser("expired-post-user", "manager", codeHash, expiry);
-
-      const response = await handleRequest(
-        mockFormRequest("/join/expired-post-123", {
-          csrf_token: "fake",
-          password: "pass12345678",
-          password_confirm: "pass12345678",
-        }),
-      );
-      expect(response.status).toBe(410);
-    });
-  });
-
-  describe("join flow (CSRF validation)", () => {
-    test("POST /join/:code rejects invalid CSRF token", async () => {
-      const { inviteCode } = await createTestInvite("csrf-test-user");
-
-      // POST with invalid (unsigned) CSRF token
-      const response = await handleRequest(
-        mockFormRequest(`/join/${inviteCode}`, {
-          csrf_token: "wrong",
-          password: "newpassword123",
-          password_confirm: "newpassword123",
-        }),
-      );
-      expectRedirectWithFlash(
-        `/join/${inviteCode}`,
-        expect.stringContaining("try again"),
-        false,
-      )(response);
-    });
-
-    test("POST /join/:code rejects missing CSRF token", async () => {
-      const { inviteCode } = await createTestInvite("csrf-missing-user");
-
-      // POST with fake CSRF token (not properly signed)
-      const response = await handleRequest(
-        mockFormRequest(`/join/${inviteCode}`, {
-          csrf_token: "token",
-          password: "newpassword123",
-          password_confirm: "newpassword123",
-        }),
-      );
-      expect(response.status).toBe(302);
-    });
-
-    test("POST /join/:code rejects form without csrf_token field", async () => {
-      const { inviteCode } = await createTestInvite("csrf-nofield-user");
-
-      // POST without csrf_token field in form
-      const body = "password=newpassword123&password_confirm=newpassword123";
-      const response = await handleRequest(
-        new Request(`http://localhost/join/${inviteCode}`, {
-          body,
-          headers: {
-            "content-type": "application/x-www-form-urlencoded",
-            host: "localhost",
-          },
-          method: "POST",
-        }),
-      );
-      expect(response.status).toBe(302);
-    });
-
-    test("POST /join/:code rejects missing password fields", async () => {
-      const { inviteCode } = await createTestInvite("validation-user");
-
-      const response = await submitJoinForm(inviteCode, {
-        password: "",
-        password_confirm: "",
-      });
-      expect(response.status).toBe(302);
-    });
-  });
-
-  describe("users template rendering", () => {
-    test("shows Invited status for user without password", async () => {
-      // Create invited user (no password yet)
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "invited-only",
-      });
-
-      const response = await awaitTestRequest("/admin/users", {
-        cookie: await testCookie(),
-      });
-      const html = await response.text();
-      expect(html).toContain("Invited");
-    });
-
-    test("shows Invite Expired status for expired invite", async () => {
-      // Create an invited user then manually set expiry to the past
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "expired-display",
-      });
-
-      const expiredExpiry = await encrypt(
-        new Date(Date.now() - 1000).toISOString(),
-      );
-      await getDb().execute({
-        args: [expiredExpiry],
-        sql: "UPDATE users SET invite_expiry = ? WHERE id = 2",
-      });
-      invalidateUsersCache();
-
-      const response = await awaitTestRequest("/admin/users", {
-        cookie: await testCookie(),
-      });
-      const html = await response.text();
-      expect(html).toContain("Invite Expired");
-    });
-
-    test("shows Pending Activation status and Activate button for user with password but no data key", async () => {
-      const { inviteCode, cookie } = await createTestInvite("pending-user");
-
-      await submitJoinForm(inviteCode, {
-        password: "newpassword123",
-        password_confirm: "newpassword123",
-      });
-
-      // Users page should show "Pending Activation" and "Activate" button
-      const usersResponse = await awaitTestRequest("/admin/users", { cookie });
-      const html = await usersResponse.text();
-      expect(html).toContain("Pending Activation");
-      expect(html).toContain("Activate");
-    });
-  });
-
-  describe("db/users.ts edge cases", () => {
-    test("verifyUserPassword returns null when user has empty password_hash", async () => {
-      const user = await createInvitedUser(
-        "nopwd",
-        "manager",
-        "hash",
-        new Date(Date.now() + 86400000).toISOString(),
-      );
-      const result = await verifyUserPassword(user, "anypassword");
-      expect(result).toBeNull();
-    });
-
-    test("isInviteValid returns false when invite_code_hash is null", async () => {
-      const user = await getUserByUsername(TEST_ADMIN_USERNAME);
-      // The owner user has no invite_code_hash
-      const valid = await isInviteValid(user!);
-      expect(valid).toBe(false);
-    });
-
-    test("isInviteValid returns false when invite was already used (empty decrypted hash)", async () => {
-      const { setUserPassword: setUserPwd } = await import("#lib/db/users.ts");
-      const expiry = new Date(Date.now() + 86400000).toISOString();
-      const user = await createInvitedUser(
-        "used-invite",
-        "manager",
-        "somehash",
-        expiry,
-      );
-
-      // Setting password clears invite_code_hash (sets to encrypted "")
-      await setUserPwd(user.id, "newpassword123");
-
-      // Reload user
-      const { getUserById: getUser } = await import("#lib/db/users.ts");
-      const updatedUser = await getUser(user.id);
-      const valid = await isInviteValid(updatedUser!);
-      expect(valid).toBe(false);
-    });
-
-    test("hasPassword returns false for user with empty encrypted password", async () => {
-      const user = await createInvitedUser(
-        "nopwd2",
-        "manager",
-        "hash2",
-        new Date(Date.now() + 86400000).toISOString(),
-      );
-      // User was created with empty password_hash - the createInvitedUser passes ""
-      const hasPwd = await hasPassword(user);
-      expect(hasPwd).toBe(false);
-    });
-
-    test("isInviteValid returns false when invite_expiry is null", async () => {
-      const { hmacHash } = await import("#lib/crypto/hashing.ts");
-      const usernameIdx = await hmacHash("no-expiry-user");
-      await getDb().execute(
-        insert("users", {
-          admin_level: await encrypt("manager"),
-          invite_code_hash: await encrypt("somehash"),
-          invite_expiry: null,
-          password_hash: "",
-          username_hash: await encrypt("no-expiry-user"),
-          username_index: usernameIdx,
-          wrapped_data_key: null,
-        }),
-      );
-      invalidateUsersCache();
-
-      const user = await getUserByUsername("no-expiry-user");
-      const valid = await isInviteValid(user!);
-      expect(valid).toBe(false);
-    });
-
-    test("decryptAdminLevel throws when admin_level decrypts to invalid value", async () => {
-      const { hmacHash } = await import("#lib/crypto/hashing.ts");
-      const usernameIdx = await hmacHash("badlevel-user");
-      await getDb().execute(
-        insert("users", {
-          admin_level: await encrypt("superadmin"),
-          invite_code_hash: null,
-          invite_expiry: null,
-          password_hash: "",
-          username_hash: await encrypt("badlevel-user"),
-          username_index: usernameIdx,
-          wrapped_data_key: null,
-        }),
-      );
-      invalidateUsersCache();
-
-      const user = await getUserByUsername("badlevel-user");
-      await expect(decryptAdminLevel(user!)).rejects.toThrow(
-        "Invalid admin level",
-      );
-    });
-
-    test("isInviteValid returns false when invite_expiry decrypts to empty string", async () => {
-      const { hmacHash } = await import("#lib/crypto/hashing.ts");
-      const usernameIdx = await hmacHash("empty-expiry-user");
-      await getDb().execute(
-        insert("users", {
-          admin_level: await encrypt("manager"),
-          invite_code_hash: await encrypt("somehash"),
-          invite_expiry: await encrypt(""),
-          password_hash: "",
-          username_hash: await encrypt("empty-expiry-user"),
-          username_index: usernameIdx,
-          wrapped_data_key: null,
-        }),
-      );
-      invalidateUsersCache();
-
-      const user = await getUserByUsername("empty-expiry-user");
-      const valid = await isInviteValid(user!);
-      expect(valid).toBe(false);
-    });
-  });
-
-  describe("utils.ts edge cases", () => {
-    test("withOwnerAuthForm returns 403 for manager role", async () => {
-      // Create manager user
-      const { hmacHash } = await import("#lib/crypto/hashing.ts");
-      const managerIdx = await hmacHash("formmanager");
-      await getDb().execute(
-        insert("users", {
-          admin_level: await encrypt("manager"),
-          password_hash: "",
-          username_hash: await encrypt("formmanager"),
-          username_index: managerIdx,
-          wrapped_data_key: null,
-        }),
-      );
-      invalidateUsersCache();
-
-      await createSession(
-        "mgr-form-session",
-        "mgr-form-csrf",
-        Date.now() + 3600000,
-        null,
-        2,
-      );
-
-      // Manager trying to POST to owner-only settings endpoint
-      // Must use a signed CSRF token so verification passes and the role check is reached
-      const { signCsrfToken } = await import("#lib/csrf.ts");
-      const signedCsrf = await signCsrfToken();
-      const response = await handleRequest(
-        mockFormRequest(
-          "/admin/settings",
-          {
-            csrf_token: signedCsrf,
-            current_password: "test",
-            new_password: "newpassword123",
-            new_password_confirm: "newpassword123",
-          },
-          `${getSessionCookieName()}=mgr-form-session`,
-        ),
-      );
-      expect(response.status).toBe(403);
-    });
-
-    test("withAuth JSON + owner role returns 403 JSON for manager", async () => {
-      // Use the manager session from createTestManagerSession
-      const cookie = await createTestManagerSession(
-        "mgr-json-session",
-        "jsonmanager",
-      );
-      const { signCsrfToken } = await import("#lib/csrf.ts");
-      // Sign CSRF inside handleRequest context via a GET that sets up the session
-      const setupResponse = await handleRequest(
-        mockRequest("/admin/", {
-          headers: { cookie },
-        }),
-      );
-      expect(setupResponse.status).toBe(200);
-
-      const { withAuth } = await import("#routes/auth.ts");
-      const { jsonResponse } = await import("#routes/response.ts");
-      const { runWithSessionContext } = await import("#lib/session-context.ts");
-
-      const response = await runWithSessionContext(async () => {
-        const signedCsrf = await signCsrfToken();
-        return withAuth(
-          mockRequest("/test", {
-            headers: {
-              "content-type": "application/json",
-              cookie,
-              "x-csrf-token": signedCsrf,
-            },
-            method: "POST",
-          }),
-          { body: "json", role: "owner" },
-          () => jsonResponse({ status: "ok" }),
-        );
-      });
-      expect(response.status).toBe(403);
-      const body = await response.json();
-      expect(body.message).toBe("Forbidden");
-    });
-  });
-
-  describe("fields.ts (username validation)", () => {
-    test("validateUsername rejects short username", async () => {
-      const { validateUsername } = await import("#templates/fields.ts");
-      expect(validateUsername("a")).toBe(
-        "Username must be at least 2 characters",
-      );
-    });
-
-    test("validateUsername rejects long username", async () => {
-      const { validateUsername } = await import("#templates/fields.ts");
-      expect(validateUsername("a".repeat(33))).toBe(
-        "Username must be 32 characters or fewer",
-      );
-    });
-
-    test("validateUsername rejects special characters", async () => {
-      const { validateUsername } = await import("#templates/fields.ts");
-      const result = validateUsername("user name");
-      expect(result).toContain("letters, numbers");
-    });
-
-    test("validateUsername accepts valid username", async () => {
-      const { validateUsername } = await import("#templates/fields.ts");
-      expect(validateUsername("valid_user-1")).toBeNull();
-    });
-  });
-
-  describe("session with deleted user", () => {
-    test("session with nonexistent user_id is rejected", async () => {
-      // Create a session pointing to a nonexistent user
-      await createSession(
-        "orphan-session",
-        "orphan-csrf",
-        Date.now() + 3600000,
-        null,
-        999,
-      );
-
-      const response = await awaitTestRequest("/admin/", {
-        cookie: `${getSessionCookieName()}=orphan-session`,
-      });
-      const html = await response.text();
-      expect(html).toContain("Login");
-    });
-  });
-
-  describe("settings updateUserPassword failure", () => {
-    test("password change returns 500 when wrapped_data_key is corrupted", async () => {
-      // Corrupt the owner's wrapped_data_key (password verification will still pass, but unwrap will fail)
-      await getDb().execute({
-        args: [],
-        sql: "UPDATE users SET wrapped_data_key = 'corrupted_key' WHERE id = 1",
-      });
-      invalidateUsersCache();
-
-      const { response } = await adminFormPost("/admin/settings", {
-        current_password: TEST_ADMIN_PASSWORD,
-        new_password: "newpassword123",
-        new_password_confirm: "newpassword123",
-      });
-      expectRedirectWithFlash(
-        "/admin/settings?form=settings-password#settings-password",
-        expect.stringContaining("Failed to update password"),
-        false,
-      )(response);
-    });
-  });
-
-  describe("settings user not found", () => {
-    test("password change returns 500 when user is deleted mid-request", async () => {
-      // Delete the user while session still exists
-      await getDb().execute({
-        args: [],
-        sql: "DELETE FROM users WHERE id = 1",
-      });
-      invalidateUsersCache();
-
-      const { response } = await adminFormPost("/admin/settings", {
-        current_password: TEST_ADMIN_PASSWORD,
-        new_password: "newpassword123",
-        new_password_confirm: "newpassword123",
-      });
-      // Session auth check should redirect since user was deleted
-      expect(response.status).toBe(302);
-    });
-  });
-
-  describe("audit logging", () => {
-    test("logs activity when user is invited", async () => {
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "auditinvite",
-      });
-
-      const logs = await getAllActivityLog();
-      expect(
-        logs.some((l) =>
-          l.message.includes("User 'auditinvite' invited as manager"),
-        ),
-      ).toBe(true);
-    });
-
-    test("logs activity when user is activated", async () => {
-      const { inviteCode, cookie, csrfToken } =
-        await createTestInvite("auditactivate");
-
-      await submitJoinForm(inviteCode, {
-        password: "newpassword123",
-        password_confirm: "newpassword123",
-      });
-
-      // Activate
-      await handleRequest(
-        mockFormRequest(
-          "/admin/users/2/activate",
-          { csrf_token: csrfToken },
-          cookie,
-        ),
-      );
-
-      const logs = await getAllActivityLog();
-      expect(
-        logs.some((l) => l.message.includes("User 'auditactivate' activated")),
-      ).toBe(true);
-    });
-
-    test("logs activity when user is deleted", async () => {
-      await adminFormPost("/admin/users", {
-        admin_level: "manager",
-        username: "auditdelete",
-      });
-
-      await adminFormPost("/admin/users/2/delete", {
-        confirm_identifier: "auditdelete",
-      });
-
-      const logs = await getAllActivityLog();
-      expect(
-        logs.some((l) => l.message.includes("User 'auditdelete' deleted")),
-      ).toBe(true);
-    });
-  });
-
-  describe("join.ts edge cases", () => {
-    test("POST /join/:code withValidInvite extracts code from params", async () => {
-      // POST to nonexistent invite code
-      const response = await handleRequest(
-        mockFormRequest("/join/nonexistent-code", {
-          csrf_token: "csrf",
-          password: "testpass123",
-          password_confirm: "testpass123",
-        }),
-      );
-      expect(response.status).toBe(404);
-    });
-  });
-});
